Rename ducks tests folder
