+++ packages/opensrp-form-config/src/components/DraftFiles/tests/index.test.tsx
@@ -8,7 +8,7 @@ import { Router } from 'react-router';
 import reducerRegistry, { store } from '@onaio/redux-reducer-registry';
 import flushPromises from 'flush-promises';
 import draftReducer, { fetchManifestDraftFiles, draftReducerName } from '../../../ducks/manifestDraftFiles';
-import { FixManifestDraftFiles, downloadFile } from '../../../ducks/tests.ts/fixtures';
+import { FixManifestDraftFiles, downloadFile } from '../../../ducks/tests/fixtures';
 import toJson from 'enzyme-to-json';
 import * as helpers from '../../../helpers/fileDownload';
 import _ from 'lodash';
