+++ packages/opensrp-form-config/src/components/Releases/tests/index.test.tsx
@@ -8,7 +8,7 @@ import { Router } from 'react-router';
 import reducerRegistry, { store } from '@onaio/redux-reducer-registry';
 import flushPromises from 'flush-promises';
 import releasesReducer, { fetchManifestReleases, releasesReducerName } from '../../../ducks/manifestReleases';
-import { fixManifestReleases } from '../../../ducks/tests.ts/fixtures';
+import { fixManifestReleases } from '../../../ducks/tests/fixtures';
 import toJson from 'enzyme-to-json';
 import _ from 'lodash';
